# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Travel API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /getFlight:
    x-swagger-router-controller: get_flight
    get:
      tags:
        - Flights
      description: Returns Flight data between the departure and arrival cities.
      operationId: getFlight
      produces:
          - "application/json"
      parameters:
          - name: departure_city
            description: Name of city from which flight should depart
            required: true
            in: query
            type: string
          - name: arrival_city
            description: Name of  city where the flight should land
            required: true
            in: query
            type: string
            
            
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/addFlightRequest"
...
definitions:
...
  addFlightRequest:
    type: object
    properties:
      airline_name:
        type: string
        description: Name of Airline
      carrier_code:
        type: string
        description: Carrier code of the Airline
      flight_number:
        type: number
        description: Flight Number
      departure_city:
        type: string
        description: Name of City where flight departs from
      arrival_city:
        type: string
        description: Name of City where the flight lands.
      departure_time:
        type: number
        description: Time in ms when the flight departs
      arrival_time:
        type: number
        description: Time in ms when the flight lands
      ticket_price:
        type: number
        description: Cost of Flight Ticket
...